feat: add board editing shortcuts to user and context menus
