+++ apps/docs/docs/management/boards/index.mdx
@@ -46,8 +46,8 @@ scrolls and store their contents separately for each layout.
 
 ## Edit a board
 
-Enter edit mode from the board header to add, move, resize, configure, or remove content. Save the board to persist the
-layout.
+Enter edit mode from the board header or account menu to add, move, resize, configure, or remove content. The account
+menu also works when the header is hidden, using the floating user button. Choose **Save board** to persist changes.
 
 - Drag from an item's inactive content area; Containers use their grip.
 - Hold `Ctrl` on Windows/Linux or `Cmd` on macOS to select several items.
@@ -56,6 +56,9 @@ layout.
 - Press Escape during a pointer or touch interaction to restore the item to its previous position.
 
 The add-content menu can create an app, integration, widget, or Container without leaving the board.
+Right-click empty board space to open this menu at the cursor. Selecting an action enters edit mode automatically.
+These actions require board editing permission; creating integrations also requires integration creation permission.
+Right-clicking existing content keeps its own menu. Hold Shift while right-clicking to use the browser menu.
 When adding apps, click one app to add it immediately, or choose **Select multiple** to add several at once. `Ctrl`/`Cmd`
 click remains available as a shortcut.
