+++ apps/nextjs/src/app/[locale]/boards/(content)/_editing-provider.tsx
@@ -1,0 +1,247 @@
+"use client";
+
+import type { PropsWithChildren } from "react";
+import { createContext, useContext, useCallback, useEffect, useRef, useState } from "react";
+import { useRouter, useSearchParams } from "next/navigation";
+import { useHotkeys } from "@mantine/hooks";
+import { notifications } from "@mantine/notifications";
+
+import { clientApi } from "@homarr/api/client";
+import { useRequiredBoard } from "@homarr/boards/context";
+import { useEditMode } from "@homarr/boards/edit-mode";
+import { revalidatePathActionAsync } from "@homarr/common/client";
+import { env } from "@homarr/common/env";
+import { hotkeys } from "@homarr/definitions";
+import { useConfirmModal } from "@homarr/modals";
+import { showErrorNotification, showSuccessNotification } from "@homarr/notifications";
+import { useI18n } from "@homarr/translation/client";
+import { useBoardPermissions } from "~/components/board/permissions/client";
+import { loadGridEditorAsync, scheduleGridEditorWarmup } from "~/components/board/sections/grid/grid-editor-loader";
+import type * as EditActionsModule from "./_edit-actions";
+
+let editActionsModulePromise: Promise<typeof EditActionsModule> | undefined;
+const loadEditActionsAsync = () => {
+  editActionsModulePromise ??= import("./_edit-actions").catch((error: unknown) => {
+    editActionsModulePromise = undefined;
+    throw error;
+  });
+  return editActionsModulePromise;
+};
+const loadBoardEditorAsync = async () =>
+  await Promise.all([
+    loadGridEditorAsync(),
+    loadEditActionsAsync(),
+    import("~/components/board/items/item-menu"),
+    import("~/components/board/sections/container/container-menu"),
+  ]);
+
+const BoardEditingContext = createContext<ReturnType<typeof useBoardEditingController> | null>(null);
+export const useOptionalBoardEditing = () => useContext(BoardEditingContext);
+
+export const BoardEditingProvider = ({
+  children,
+  enabled,
+  demoReadOnly,
+}: PropsWithChildren<{ enabled: boolean; demoReadOnly: boolean }>) => {
+  if (!enabled) return children;
+  return <ActiveBoardEditingProvider demoReadOnly={demoReadOnly}>{children}</ActiveBoardEditingProvider>;
+};
+const ActiveBoardEditingProvider = ({ children, demoReadOnly }: PropsWithChildren<{ demoReadOnly: boolean }>) => {
+  const controller = useBoardEditingController(demoReadOnly);
+  return <BoardEditingContext.Provider value={controller}>{children}</BoardEditingContext.Provider>;
+};
+const useBoardEditingController = (demoReadOnly: boolean) => {
+  const [isEditMode, { open, close }] = useEditMode();
+  const [editorLoadState, setEditorLoadState] = useState<"scheduled" | "loading" | "ready" | "error">("scheduled");
+  const [isEnteringEditMode, setIsEnteringEditMode] = useState(false);
+  const board = useRequiredBoard();
+  const { hasChangeAccess } = useBoardPermissions(board);
+  const utils = clientApi.useUtils();
+  const router = useRouter();
+  const searchParams = useSearchParams();
+  const { openConfirmModal } = useConfirmModal();
+  const tCommon = useI18n("common");
+  const tBoardSetting = useI18n("board.setting");
+  const t = useI18n("board.action.edit");
+  const returnToLayoutSettings =
+    isEditMode &&
+    searchParams.get("edit") === "true" &&
+    searchParams.get("layout") !== null &&
+    searchParams.get("returnTo") === "settings";
+  const latestBoardRef = useRef(board);
+
+  latestBoardRef.current = board;
+  const { mutate: saveBoard, isPending } = clientApi.board.saveBoard.useMutation({
+    onMutate(boardToSave) {
+      boardToSave.items.forEach(({ id }) => notifications.hide(`board-item-created:${id}`));
+    },
+    onSuccess() {
+      showSuccessNotification({
+        title: tCommon("notification.update.success"),
+        message: t("notification.success.message"),
+      });
+      void utils.board.getBoardByName.invalidate({ name: board.name });
+      void utils.widget.customApi.getData.invalidate();
+      void revalidatePathActionAsync(`/boards/${board.name}`);
+      close();
+      if (returnToLayoutSettings) router.push(`/boards/${board.name}/settings?tab=layout`);
+    },
+    onError() {
+      showErrorNotification({
+        title: tCommon("notification.update.error"),
+        message: t("notification.error.message"),
+      });
+    },
+  });
+
+  const discardDemoChanges = useCallback(() => {
+    void utils.board.getBoardByName.invalidate({ name: board.name });
+    close();
+  }, [utils, board.name, close]);
+
+  const prepareEditorAsync = useCallback(async () => {
+    if (editorLoadState === "ready") return;
+
+    setEditorLoadState("loading");
+    try {
+      await loadBoardEditorAsync();
+      setEditorLoadState("ready");
+    } catch {
+      setEditorLoadState("error");
+      throw new Error("Unable to load the board editor");
+    }
+  }, [editorLoadState]);
+
+  useEffect(() => {
+    return scheduleGridEditorWarmup(async () => {
+      setEditorLoadState("loading");
+      try {
+        await loadBoardEditorAsync();
+        setEditorLoadState("ready");
+      } catch {
+        setEditorLoadState("error");
+      }
+    });
+  }, []);
+
+  const prewarmEditor = useCallback(() => {
+    void prepareEditorAsync().catch(() => {
+      // A click or hotkey retries and reports the error.
+    });
+  }, [prepareEditorAsync]);
+
+  const enter = useCallback(async () => {
+    if (!hasChangeAccess || isPending || isEnteringEditMode) return false;
+    if (isEditMode) return true;
+    setIsEnteringEditMode(true);
+    try {
+      await prepareEditorAsync();
+      open();
+      return true;
+    } catch {
+      showErrorNotification({
+        title: t("notification.loadError.title"),
+        message: t("notification.loadError.message"),
+      });
+      return false;
+    } finally {
+      setIsEnteringEditMode(false);
+    }
+  }, [hasChangeAccess, isPending, isEnteringEditMode, isEditMode, open, prepareEditorAsync, t]);
+
+  const toggle = useCallback(async () => {
+    if (!hasChangeAccess || isPending || isEnteringEditMode) return;
+    if (!isEditMode) return enter();
+    if (demoReadOnly) return discardDemoChanges();
+    saveBoard(latestBoardRef.current);
+  }, [hasChangeAccess, isPending, isEnteringEditMode, isEditMode, enter, demoReadOnly, discardDemoChanges, saveBoard]);
+
+  const cancelLayoutEdit = useCallback(() => {
+    openConfirmModal({
+      title: tBoardSetting("section.layout.edit.cancel.title"),
+      children: tBoardSetting("section.layout.edit.cancel.message"),
+      confirmProps: { children: tCommon("action.discard"), color: "red" },
+      onConfirm() {
+        void (async () => {
+          await utils.board.getBoardByName.invalidate({ name: board.name });
+          close();
+          router.push(`/boards/${board.name}/settings?tab=layout`);
+        })();
+      },
+    });
+  }, [board.name, close, openConfirmModal, router, tBoardSetting, tCommon, utils.board.getBoardByName]);
+
+  useHotkeys([[hotkeys.toggleBoardEdit, () => void toggle()]]);
+  usePreventLeaveWithDirty(isEditMode && hasChangeAccess);
+
+  let label = t("enter");
+  if (isEditMode) label = t("save");
+  if (isEditMode && demoReadOnly) label = t("exit");
+  return {
+    isEditMode,
+    hasChangeAccess,
+    label,
+    toggle,
+    enter,
+    prewarmEditor,
+    editorLoadState,
+    isPending,
+    isEnteringEditMode,
+    returnToLayoutSettings,
+    cancelLayoutEdit,
+  };
+};
+
+const anchorSelector = "a[href]:not([target='_blank'])";
+const usePreventLeaveWithDirty = (isDirty: boolean) => {
+  const tBoard = useI18n("board");
+  const tCommon = useI18n("common");
+  const { openConfirmModal } = useConfirmModal();
+  const router = useRouter();
+
+  useEffect(() => {
+    if (!isDirty) return;
+
+    const handleClick = (event: MouseEvent) => {
+      if (event.button !== 0 || event.metaKey || event.ctrlKey || event.shiftKey || event.altKey) return;
+      const target = event.target instanceof Element ? event.target.closest<HTMLAnchorElement>(anchorSelector) : null;
+      if (!target) return;
+
+      event.preventDefault();
+      event.stopPropagation();
+
+      openConfirmModal({
+        title: tBoard("action.edit.confirmLeave.title"),
+        children: tBoard("action.edit.confirmLeave.message"),
+        onConfirm() {
+          router.push(target.href);
+        },
+        confirmProps: {
+          children: tCommon("action.discard"),
+        },
+      });
+    };
+
+    const handlePopState = (event: Event) => {
+      window.history.pushState(null, document.title, window.location.href);
+      event.preventDefault();
+    };
+
+    const handleBeforeUnload = (event: BeforeUnloadEvent) => {
+      if (env.NODE_ENV === "development") return;
+
+      event.preventDefault();
+      event.returnValue = true;
+    };
+
+    document.addEventListener("click", handleClick, true);
+    window.addEventListener("popstate", handlePopState);
+    window.addEventListener("beforeunload", handleBeforeUnload);
+
+    return () => {
+      document.removeEventListener("click", handleClick, true);
+      window.removeEventListener("popstate", handlePopState);
+      window.removeEventListener("beforeunload", handleBeforeUnload);
+    };
+  }, [isDirty, openConfirmModal, router, tBoard, tCommon]);
+};
